(v0.22.0) Add host class package to anon class name for 11+
Reflection fails for anonymous classes with a host class in a module which
exports only certain packages, since in OpenJ9, the anonymous class name does
not include package info. In the reference implementation, anonymous classes
do have package info in their name, so this commit changes OpenJ9 behaviour to
match that.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>