Recognize Unsafe.copyMemory0 in JDK11
OpenJ9 contains the optimization that recognizes the JNI call
sun/misc/Unsafe.copyMemory and transforms it to System.arrayCopy calls,
which is inlined and optimized on most of the platform. Due to
implementation changes in JDK11 we can not apply this optimization to
sun/misc/Unsafe.copyMemory method which in Java 11 acts as a java
wrapper that calls the implementation of copyMemory method from
jdk/internal/misc/Unsafe class. This implementation contains additional
changes to check range for the source and destination (in case of
illegal access throws RuntimeException) before calling the
actual JNI method jdk/internal/misc/Unsafe.copyMemory0 which can be
transformed to System.arrayCopy call. Due to this behavioural changes,
when compiling method that calls Unsafe.copyMemory, we need to make sure
that optimizer is exposed to JNI call Unsafe.copyMemory0 and that call
is recognized same way as JNI call sun/misc/Unsafe.copyMemory is
recognized in Java 8 to optimize it further.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>