Fix incorrect BNDCHK immediate bound check under -Xrs
Previously, there was a bug in the Power code generator's implementation
of BNDCHK where it would try to encode small negative integers as the
immediate of a cmplwi instruction when traps were disabled. This was
caused by the bounds checks being shared between the traps enabled case,
which uses the twi instruction (which expects a signed immediate), and
the traps disabled case, which uses the cmplwi instruction (which
expects an unsigned immediate).

For the time being, this has been fixed by adjusting the bounds checks
on the immediate to exclude small negative values. These should be
relatively rare, as BNDCHK nodes with negative constant children should
always throw.

Signed-off-by: Ben Thomas <ben@benthomas.ca>