Correctly set invocationCount in Valhalla tests
In order to ensure the JIT is adequately exercised, some tests have to
be run at least twice so that more compilations of key methods are
triggered. Previously, the Valhalla tests attempted to do this by
passing `invocationCount=2` on the @Test annotation on the test class.
However, TestNG seems to only handle that parameter when passed to the
@Test annotation on a method. This commit removes the invocationCount
parameter from the class annotation and adds it to individual test
methods that can be safely invoked more than once. Certain tests cannot
be safely run more than once because they are stateful. In this commit,
only tests that are obviously safe are set to be invoked twice. Other
tests will require closer investigation and possibly some refactoring
and are not addressed in this commit to avoid delaying the progression
of other work.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>