Introduce instanceHasRefs flag
There is an issue with instance description bits with respect to
flattened types. See https://github.com/eclipse/openj9/issues/10070.
Currently, there is a mistach between the layout of a type on its own,
and when it is flattened. This is due to the pre-padding required for
standalone valuetypes on compressedrefs mode. The `backfill` has been
used to track pre-padding. This is an error prone approach and has cause
bugs in the past.

The only cases where pre-padding is needed is on compressedRefs where
there is at least one double-slot sized field. A better solution (to
backfill approach) is, instead of pre-padding, insert a single slot
field in the spot where the padding would be. This this means that if a
type (in compressedrefs mode) has at least one single slot or ref field
the pre-padding is not needed. This leaves fields with only double slot
fields. In these cases, we do not need to track the padding because
double-slot fields cannot be references, so there is no mismatch with
the instance description bits between the standalone layout and the
flattened layout.

This PR is the first step in solving this problem. This PR introduces a
flag on the J9Class that indicates whether the type has instances field
which are references. In these cases the barrier API for copying
instance fields doesn't need to scan the instance description bits to
apply barrier loads/stores. It can simply use a slot sized load/store
for the operation.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>