Check for NULL thread lock
When suspending a thread, handle the case where the Thread lock object
has not yet been assigned. This can occur if an event (e.g. method
enter) is reported during the construction of the Thread object. The
thread will be in the thread list, but the object will not be fully
intitialized.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>