Add new TR_RelocationRecord::create API
Add a new API that does not require the relocation record binary
template header to already exist in order to determine the kind of the
relocation record when creating it. This is necessary because the
TR_RelocationRecord and subclasses are an API class that are used to
both read and write data to and from, respectively, the binary template
header.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>