Introduce the !showflags jdmpview command
showflags will output the user's specified flag structure and the corresponding values of each flag.

There are 3 method of running this command.
Method 1: !showflags
When running just the !showflags command, it will output all of the possible flag structures that are available

Method 2: !showflags flag_structure
When specifying a certain flag structure, the constants of that flag structure will be outputted (along with all of their values)

Method 3: !showflags flag_structure specific_flag
When specifiying a certain flag, along with the structure that it belongs to, the corresponding value will be outputted

Signed-off-by: Dusan-Boskovic <dusan.boskovic@ibm.com>