Consolidate TR_MethodObject
Remove the duplicated code that writes the
TR_MethodObject relocation header
information, and consolidate it in one canonical location.

Note: TR_MethodObject is not used anywhere on P; in fact that code that
wrote out the header data was incorrect as it wrote in the fields in the
wrong order.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>