Maintain CH table per JITServer client instead of globally
Before this commit, server had one global CH table access
to which was controlled by one global lock.
This is not ideal when multiple clients connect to the server
because it leads to scalability issues due to lock contention
and also one client might retrieve data that belongs to another client.

This commit replaces the global CH table with one created per client.
Lock contention is also removed as we now have one lock per client as
well.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>