Delete duplicate table of sizes
Update the code to only use the canonical table of relocation record
header sizes in RelocationRecord.cpp, rather than the duplicated (for
each platform) table of sizes. Additionally, delete these now unused
tables.

Additionally, this commit also deletes code that guarded by #ifdef 0
in ARM for two reasons; 1, it's #ifdef 0, and 2, it's code that prints
out the fields in the relo header which is done elsewhere.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>