Exploit prefixed instruction in IPIC sequence on POWER10
Previously, either pTOC entry or long sequence are used in gettting
the snippet address. Prefixed instruction can be used for the same
purpose nicely on P10 without pTOC.

Correspondingly, interface snippet emitting needs to change. In
particular, CallSitePatching needs to be adjusted as well.

Type-casting for the top 18-bit offset to avoid potential loss of bits.

Split the code to make it clearer.

Signed-off-by: Julian Wang <zlwang@ca.ibm.com>