Rewrite MH.invoke and MH.invokeExact to invokehandle for OpenJDK MHs
For OpenJDK MethodHandles (MHs), MH.invoke and MH.invokeExact have the
same behavior. So, MH.invoke and MH.invokeExact are translated to the
same bytecode, "invokehandle".

The OpenJDK MH implementation relies upon LambdaForm(s) to dynamically
generate bytecodes on the Java side. So, operations such as asType are
handled on the Java side. This allows MH.invoke/invokeExact to share a
common path in the interpreter.

For OpenJ9 MHs, MH.invoke is translated to the invokehandlegeneric
bytecode, and MH.invokeExact is translated to the invokehandle bytecode.

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>