Rewrite VarHandle polymorphic methods to invokehandle bytecode
For the OpenJDK implementation, the VarHandle polymorphic methods are
translated to the invokehandle bytecode.

The OpenJDK MH/VH implementation relies upon LambdaForm(s) to
dynamically generate bytecodes on the Java side. So, operations such as
asType are handled on the Java side. This allows MH.invoke,
MH.invokeExact and VH polymorphic methods to share a common path in the
interpreter.

For the OpenJ9 implementation, the VarHandle polymorphic methods are
redirected to the invokevarhandle INL during resolution.

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>