Remove the un-needed NOP added for adjusting GCMap
For Unresolved Virtual Call Dispatch sequence in read only code, we had
to add a NOP in mainline before the start of the resolve sequence for
such calls to make sure that if a GC is triggered while calling C helper
function, it sees correct GCMap. This was a hack with potential bug where
local register assigner can shuffles the register between NOP and start
of virtual call resolution sequence which can alter the state of
registers.
Correct fix for this issue is implemented in this commit where while
calling C helper from PicBuilder, it passes the address of the next
instruction in mainline after snippet call.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>