Fix GCMap issue in Unresolved field Resolution
For the unresolved field resolution, JIT helper glue code will have set
up the return address to be the start of the resolution sequence where
after returning from the helper, we again load the correct index or
address from the resolution data block. If the helper we call ends up
triggering GC, the stack walker will use the JIT return address we
passed to walk current JIT frame and find the GC map associated with the
return address. The GC map logic subtracts 1 which means that unless we
create the GC Map on an instruction before the return label, we will
retrive the wrong GC map. Correct fix here would be passing the address
of next instruction after the snippet call in mainline as jitEIP when
we call JIT helper. This commit fixes this issue.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>