Check whether TLS storage for comp object was created before freeing it
When the JIT fails initialization it proceeds with the
shutdown routines. In here the JIT tries to deallocate
the TLS storage for the compilation object. However,
because the JIT failed so early, this storage might
not have been allocated in the first place.
This PR checks whether the TLS storage was allocated
before attempting to free it.

Depends on: https://github.com/eclipse/omr/pull/5549
Issue: #10339

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>