Convert JIT metadata offsets to SRPs
SRPs are easier to work with and more efficient, as long as
we can guarantee that the offsets are stable.

The JIT metadata lives in the data cache; references from the
metadata to anything else in the data cache will of course be
stable, but references to things outside the data cache (J9
methods, constant pools, etc.) will need to maintain their
relative offset from the data cache for SRPs to work.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>