Use a separate mutex for small freed block lists of persistent allocator
Small freed memory blocks using persistent memory are kept in some
linked lists, where each list only contains blocks of same size.
This allows the JIT to rapidly allocate/deallocate a small block
of memory by detaching the head (or adding to the head) of an
appropriate such list.
Larger freed blocks are kept in a separate list where blocks are
ordered by size. The overhead to access this list can sometimes be
quite large, depending on the number of entries in this list (we
need to traverse the list to find the right location for adding
to or deleting from it).
Currently, all these lists are protected by the same monitor:
`memoryAllocMonitor`. In order to avoid a fast allocation or
deallocation of a small block to be delayed by a concurrent access
to the large-block list (which can take time), this commit
introduces a dedicated mutex for the set of small-freed-block
lists, called `_smallBlockListsMonitor`.
Some refactoring of the code has also been performed.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>