Update travis git cache to fetch snapshot branch
travis, by default, clones the repo with:
```
git clone --depth=50
```
which implies `--single-branch`.  For the snapshot builds,
we need the snapshot brnach to be present in the local git
repo.

To make that work, we need to add the remote reference for
the snapshot branch and then fetch it with the depth of 50.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>