JITServer: Eliminate cpu.getProcessorDescription() remote override
There is only one place in the entire codebase where it makes sense to grab processor description from the client.
There is no need to do a remote override for this method unlike the other methods that get called in many places in the codebase.
Eliminate the remote call to client in getProcessorDescription to avoid any confusions and make the code a little bit faster.

Signed-off-by: Harry Yu <harryyu1994@gmail.com>