Chnage to different project assigner action
Change to use the action from:
https://github.com/marketplace/actions/assign-to-one-project

to handle assign issues and pull requests with the
`snaphot` label to the `Snapshot+Restore` project.

This has been tested on my https://github.com/DanHeidinga/ToyTest
repo and shown to work there.

Note, Github builds the docker image from the provided Dockerfile
in the action repo so we can trust the image matches the source
in that repo.
See https://docs.github.com/en/free-pro-team@latest/actions/creating-actions/creating-a-docker-container-action
for details on the docker image build process.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>