Auto download bootjdk
If bootjdk is not present on the system in the
location provided, attempt to pull one from Adopt
and unpack it to that same location.

Assumes build user has write permissions in the
location provided. Downloads version specified
in the variable file.

Also reworks function check_path to return a bool
rather than either a string or empty string.

[skip ci]
Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>