Move static fields and methods from MethodHandle to MethodHandleHelper
The code has been verbatim moved from MethodHandle to
MethodHandleHelper.

The following static final fields are moved from MethodHandle to
MethodHandleHelper:
- UNSAFE
- JITHELPERS
- BSM_ARGUMENT_SIZE
- BSM_ARGUMENT_COUNT_OFFSET
- BSM_ARGUMENTS_OFFSET
- BSM_LOOKUP_ARGUMENT_INDEX
- BSM_NAME_ARGUMENT_INDEX
- BSM_TYPE_ARGUMENT_INDEX
- BSM_OPTIONAL_ARGUMENTS_START_INDEX

The following native static methods are moved from MethodHandle to
MethodHandleHelper:
- getCPTypeAt
- getCPMethodTypeAt
- getCPMethodHandleAt
- getCPClassNameAt
- getCPConstantDynamicAt

The following Java static methods are moved from MethodHandle to
MethodHandleHelper:
- resolveConstantDynamic
- constructorPlaceHolder
- invokeBsm
- resolveInvokeDynamic
- sendResolveMethodHandle
- resolveFieldHandleHelper
- getClassFromJ9Class
- fromFieldDescriptorString
- getAdditionalBsmArg
- throwNoClassDefFoundError
- getJ9ClassFromClass

This is done to support the OpenJDK MethodHandle class. Some of the
above Java static methods will be updated to support the OpenJDK
MethodHandle class.

This will also resolve a limitation of vmconstantpool.xml, which exposes
some of the above Java static methods in the native code. The OpenJDK
MethodHandle class does not have the above Java static methods. Having
the static methods in MethodHandleHelper allows them to be exposed to
both OpenJ9 and OpenJDK implementations. This prevents the JCL
initialization crash, associated to vmconstantpool.xml, when the OpenJDK
MethodHandle class is enabled.

Related: https://github.com/eclipse/openj9/issues/7352

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>