Use flattened types as backfill
There is an issue with instance description bits with respect to
flattened types. See #10070.

Currently, there is a mistach between the layout of a type on its own,
and when it is flattened. This is due to the pre-padding required for
standalone valuetypes on compressedrefs and 32bit mode. The backfill has
been
used to track pre-padding. This is an error prone approach and has cause
bugs in the past.

The only cases where pre-padding is needed is on compressedRefs where
there is at least one double-slot sized field. A better solution (to
backfill approach) is, instead of pre-padding, insert a single slot
field in the spot where the padding would be. This this means that if a
type (in compressedrefs/32bit mode) has at least one single slot or ref
field
the pre-padding is not needed. This leaves fields with only double slot
fields. In these cases, we do not need to track the padding because
double-slot fields cannot be references, so there is no mismatch with
the instance description bits between the standalone layout and the
flattened layout.

Instead of using backfill this PR proposes an approach that makes
pre-padding explicit ie `J9CLASS_HAS_4BYTE_PREPADDING`. When a type is
flattened the pre-padding is eliminated, so offsets into flattened types
need to take this into account.

The backfill logic has been updated. The new back fill ordering is:
1. instance singles
2. instance objects
3. flat end-aligned singles
4. flat end-aligned objects
5. flat end-unaligned singles
6. flat end-unaligned objects
7. hidden singles
8. hidden objects

This ordering only applies to valuetypes, in the future this may be
extended to reference types as well.

If a flattenable type requires double alignment in the past the type
will be pre-padded. Instead we will now try to use up the pre-padding
with the fields that do not require double alignment (the precedence is
shown above). If no types are available then we will insert pre-padding
just as before. If the type used for pre-padding is sized in manner that
a double can not be placed after it contiguously then the type is
post-padded, this is no worse than adding pre-padding.

This PR also re-enables tests that previously didnt work due to issues
with field layouts.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>