Add support to resolve invokedynamic under OpenJDK MethodHandles
MethodHandleNatives.linkCallSite is used to resolve invokedynamic.

The resolved values are returned in a two element array. The first array
element is a MemberName object, and the second array element is a
MethodHandle object.

The return type of resolveInvokeDynamic is changed to Object in order to
support both the OpenJ9 and OpenJDK implementations for resolving
invokedynamic. For OpenJDK MethodHandles, the two element array is
casted to an Object before being returned. For OpenJ9 MethodHandles, the
MethodHandle object automatically qualifies as an Object. The change in
return type is reflected in vmconstantpool.xml.

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>