Add support for OpenJDK (OJDK) MethodHandle (MH) in MethodHandleResolver
1. Update sendResolveMethodHandle and resolveFieldHandleHelper for OJDK MH

For OJDK MH, the usage of accessCheckArgRetTypes has been commented from
sendResolveMethodHandle and resolveFieldHandleHelper since this method
is not available in OpenJDK Lookup.

TODO: Investigate if an equivalent for lookup.accessCheckArgRetTypes is
needed for OJDK MH. If needed, find an alternative for
lookup.accessCheckArgRetTypes.

OpenJDK Lookup does not have constructor Lookup(Class<?>, boolean). So,
it has been replaced with constructor Lookup(Class<?>) in
sendResolveMethodHandle.

2. Add support to resolve invokedynamic under OpenJDK MethodHandles

MethodHandleNatives.linkCallSite is used to resolve invokedynamic.

The resolved values are returned in a two element array. The first array
element is a MemberName object, and the second array element is a
MethodHandle object.

The return type of resolveInvokeDynamic is changed to Object in order to
support both the OpenJ9 and OpenJDK implementations for resolving
invokedynamic. For OpenJDK MethodHandles, the two element array is
casted to an Object before being returned. For OpenJ9 MethodHandles, the
MethodHandle object automatically qualifies as an Object. The change in
return type is reflected in vmconstantpool.xml.

3. Add support to resolve constantdynamic under OpenJDK MethodHandles

For OpenJDK MethodHandles, MethodHandleNatives.linkDynamicConstant is
used to resolve constantdynamic.

The resolve value is returned in the form of an Object, which is
identical to the OpenJ9 implementation.

4. Add the implementation for linkCallerMethod

linkCallerMethod was previously named as resolveMethodHandle.

It's a wrapper for MethodHandleNatives.linkMethod. It is used in
invokehandle for MethodHandle and VarHandle polymorphic methods. The
resolution yields two values. The two values are returned in a two
element array. The first array element is a MemberName object, which
specifies the caller method to be invoked. The second array element is a
MethodType object if the defining class is a MethodHandle and an
AccessDescriptor object if the defining class is a VarHandle.

linkCallerMethod is only needed for OJDK MH. InternalError will be
thrown if it is invoked for OpenJ9 MH.

5. Disable invokeBsm for OJDK MH

invokeBsm is unused for OpenJDK MethodHandles.

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>