Add the implementation for resolveMethodHandle
It's a wrapper for MethodHandleNatives.linkMethod. It is used in
invokehandle for MH.invoke and MH.invokeExact. The resolution yields two
values. One is returned and the other value is stored in an one element
array named, appendixResult. The value returned is a MemberName object,
which specifies the caller method to be invoked. The value stored in
appendixResult is a MethodType object in case the defining class is a
MethodHandle and an AccessDescriptor object in case the defining class
is a VarHandle.

resolveMethodHandle is only needed for OJDK MH.

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>