Change TR_ASSERT_FATAL into failCompilation call
`findOrCreateShadowSymbol()` tries to reuse symrefs produced by
`findOrFabricateShadowSymbol()` when the shadow corresponds to the same
field. Detecting that the fields are in fact the same requires checking
that the defining class is the same. Normally, the expectation is that
if a cp field ref is resolved, then the cp ref pointing to the defining
class must also be resolved. However, in #9416, it was discovered that
this assumption does not hold in some cases. Under AOT, loading class
data from the SCC using `findSharedData()` may fail, which is propagated
to the JIT as a NULL defining class. In addition, if a class is
redefined between the point where a field is looked up and where the
defining class is looked, the latter may become NULL.

Since the code in question will not work correctly if the defining class
is null, this commit tactically replaces the assert with a call to
`failCompilation()`.

Fixes #9416

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>