Improve performance of persistent memory allocator
The persistent memory allocator keeps a single linked list
for all freed memory blocks larger than 128 bytes. The list
is sorted by the block size so every time we need to allocate
from it or add a freed block to it we need to traverse it
to find the appropriate spot. If this list becomes very large,
searching the list for the appropriate spot could consume an
excessive amount CPU.
For a regular JVM the usage pattern is not conducive to very
large such lists. In contrast, JITServer makes extensive use
of persistent memory. When one JVMClient ends, all of its
persistent allocations (at JITServer) are put on this free
block list which could reach over 30K elements in size.
This makes allocations for a subsequent JVMClient very slow.
The solution implemented by this commit is to keep a set
of indices to nodes whose sizes are power-of-two. This way,
when searching for a free block, we don't have to start from
the beginning of the list, but start from a node in the middle
based on size.
Since this change carries a bit of risk, it will be enabled
(for the time being) only for JITServer

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>