Reduce the number of ResolvedMethod_getResolvedImplmentorMethods
This message is sent when the CH table needs to walk the class
tree starting from some top class, and find at most some specified
number of implementor methods.
This commit reduces the number of messages by caching which class
trees have already been walked during current compilation.
If the tree has been walked, all required implementor methods have
already been cached and thus we do not require another remote call.

To accomodate this behavior, I added methods to get resolved methods
only from the cache, not making a remote call if the methods is not
cached.
This guarantees that when the class tree is walked on the server,
the only non-NULL methods will be the cached implementors.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>