Remove use of reverse loads/store for DAA
Previously, some methods in the DAA library would use the reverse load
opcodes to perform a byte-swapped load and the reverse store opcodes to
perform a byte-swapped store. However, these opcodes are slated for
removal, as they can be replaced with a normal load/store and a
byteswap. This replacement has now been done to allow the removal of
these opcodes.

Signed-off-by: Ben Thomas <ben@benthomas.ca>