Remove reverse load/store opcodes
Previously, OpenJ9 has a number of opcodes for performing byteswapped
loads and stores. However, OMR has since added a family of byteswap
opcodes that can be used on load results and store values instead to get
the same effect in a much easier way. This allows support for these
operations to be extended to any codegen that supports byteswap without
the need for them to add support for new types of loads, which is much
more invasive.

As a result of this, these old opcodes are no longer needed and are thus
being removed to get rid of a lot of legacy code that is no longer
actively in use.

Signed-off-by: Ben Thomas <ben@benthomas.ca>