Consolidate code for checking a VarHandle polymorphic method
There are two locations where redundant code is used to check if a
class method is a VarHandle polymorphic method. The two locations are
the following:
1) ConstantPoolMap::isVarHandleMethod
2) ClassFileOracle::isPolymorphicVarHandleMethod

A new utility function, VM_VMHelpers::isPolymorphicVarHandleMethod, has
been added to support the above two locations.

The redundant code has been removed, and the new utility function is
being used in the above two locations.

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>