Fix incorrect helper trampolines when disableTOC is set
During a previous rework of how trampolines are generated, support was
added for helper trampolines to be generated when the pTOC is explicitly
disabled. However, this implementation accidentally used an oris
instruction where it should have used an ori instruction, which results
in the branch going to the wrong address. This bug has been fixed.

Fixes: #10939
Signed-off-by: Ben Thomas <ben@benthomas.ca>