Add new `allSegmentsInMemorySegmentListDo` api to iterate segments
The new api iterates through all the J9MemorySegments in a given
J9MemorySegmentList and passes them to a callback function.  This
allows arbitrary operations to be performed on the segment.

I've validated this with adhoc tests that print the segments and
remove bits from them.

The API is intended to help the `snapshot` work but as it's common
and can be put immediately into `master`, that's my preference to
avoid porting it over later.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>