Uncommon autosimd piv loop stride increment node
If the loop increment node is commoned, then reducing the loop by
changing the value of loop stride can cause unwanted bug. This is
potentially true if the node is commoned with address calculation.
This commit uncommon the PIV increment node so that changing the
stride does not introduce side-effects on commoned nodes. Also, the
fix looks for commoning of ifcmp checks that are commoned with the
stride calculation and adjust them accordingly.

Fixes #10656

Signed-off-by: Mohammad Nazmul Alam <mohammad.nazmul.alam@ibm.com>