Disable initializeMethodRunAddressVarHandle for OpenJDK MethodHandles
For OpenJDK VarHandles, the invokevirtual calls to the VarHandle
polymorphic methods are rewritten to the invokehandle bytecode in
ClassFileOracle (https://github.com/eclipse/openj9/pull/10557). The
invokevarhandle INL is not used with OpenJDK VarHandles. For OpenJDK
VarHandles, the VarHandleInternal class is unused, and its methods do
not have to be translated to the invokevarhandle INL. So,
initializeMethodRunAddressVarHandle is disabled for OpenJDK VarHandles.

Related: https://github.com/eclipse/openj9/issues/7352

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>