Enable OpenJDKCompileStub.OpenJDKCompileStubThrowError() in IBM builds
OpenJDKCompileStubThrowError() can be used universally, in both IBM and
OpenJ9 builds, to pin point a stub method, which should not be used due
to lack of an implementation.

MethodHandleResolver.linkCallerMethod() invokes
OpenJDKCompileStubThrowError() in IBM builds. So, its implementation is
being exposed to IBM builds for resolving a compile-time error.

Related: https://github.com/eclipse/openj9/issues/7352

[ci skip]

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>