Use new incInlineDepth API
Use the new incInlineDepth API to ensure that the inlined site contains
all the information needed for relocation in case a guard does not exist
for that site. Additionally, add an assert in the old API to ensure that
it is not called except for when the callsite is not actually
incremented (such as when estimating the bytecode size in the inliner
prior to actually inlining).

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>