Change data passed in to inlining table relocations
Because an inlined site may not have a guard, this commit changes the
data that's passed in when creating the external relocation. Previously
the virtual guard would be passed in, which would be used to acquire all
the necessary information; now the data is put into a
TR_RelocationRecordInformation struct first, and this struct is passed
into the external relocation.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>