Address code review comments
Reworked to have Walker.cpp abort compilation in most cases if OSR
cannot be induced with recompilation requested rather than always
delaying aborting until HandleRecompilationOps is run.  The conditions
that are checked have been moved into a common method,
TR::Compilation::isOSRAllowedForOperationsRequiringRecompilation.
That allows HandleRecompilationOps::perform to exit early in cases where
the tranformation it performs is absolutely not needed.

Also corrected HandleRecompilationOps::perform to return 1 if it
performs any transformation.

Finally, changed signature of visitNode to accept only a tree to make
it clear that it wouldn't be recursively called for all the nodes in
the tree, and renamed it to visitTree.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>