Deserialize runtime assumptions from buffer
This method loops through all the runtime assumption kinds. For each
kind (that's supported), it calls the appropriate static method
deserialize. The reason deserialize is static rather than virtual is
because the constructors of the runtime assumptions are protected;
therefore, we can't initialize a dummy instance to call a hypothetical
virtual deserialize.

Therefore, deserialize takes a pointer to the start of the section
that contains the deserialized assumptions for that kind as well
as the number of assumptions in that section. It then loops through
the section and calls make to create the runtime assumptions and
add it to the RAT.

TR_RuntimeAssumptionTable::deserialize also takes an optional
parameter kind, which can be used to only create assumptions of a
specific kind (rather than all kinds).

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>